# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Persona Runtime public API",
    "version": "1.0.0-draft.2",
    "description": "목표 계약 초안. 초기 제품 제한·보존·토큰 메모리 보관 반영. 서버 구현 완료 아님. 동작은 service-api-v1.md, 차이는 implementation-gap-v1.md. 남은 정책은 x-pending-policy-decisions. 합성 예시만 사용."
  },
  "servers": [
    {
      "url": "/",
      "description": "동일 origin의 Tailnet HTTPS 진입점. 개발 mock 이외 실제 토큰을 평문으로 보내지 않는다."
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "접속 토큰·현재 사용자 확인",
        "tags": [
          "Identity"
        ],
        "description": "접속 토큰·현재 사용자 확인",
        "parameters": [],
        "responses": {
          "200": {
            "description": "접속 토큰·현재 사용자 확인",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
    "/v1/personas": {
      "post": {
        "operationId": "createPersona",
        "summary": "자료 없는 캐릭터 생성",
        "tags": [
          "Personas"
        ],
        "description": "자료 없는 캐릭터 생성. 사용자당 삭제 중 포함 최대 3개, 삭제 완료 후 슬롯 반환. 한도 초과 409 persona_limit_exceeded. 동일 키 replay는 새 캐릭터를 만들지 않으며 한도 검사·등록을 원자 처리.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "자료 없는 캐릭터 생성",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Persona"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersona"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listPersonas",
        "summary": "내 캐릭터 목록",
        "tags": [
          "Personas"
        ],
        "description": "내 캐릭터 목록",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          }
        ],
        "responses": {
          "200": {
            "description": "내 캐릭터 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
    "/v1/personas/{persona_id}": {
      "get": {
        "operationId": "getPersona",
        "summary": "적용본·초안 상태 포함 상세 조회",
        "tags": [
          "Personas"
        ],
        "description": "적용본·초안 상태 포함 상세 조회",
        "parameters": [
          {
            "in": "path",
            "name": "persona_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "적용본·초안 상태 포함 상세 조회",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaDetail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      },
      "delete": {
        "operationId": "deletePersona",
        "summary": "캐릭터 비동기 삭제 접수",
        "tags": [
          "Operations"
        ],
        "description": "캐릭터 비동기 삭제 접수",
        "parameters": [
          {
            "in": "path",
            "name": "persona_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "캐릭터 비동기 삭제 접수",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    },
...
    "/v1/personas/{persona_id}/retrieve": {
      "get": {
        "operationId": "retrievePersonaContext",
        "summary": "검색 결과 디버그 조회",
        "tags": [
          "Personas"
        ],
        "description": "질문으로 본문·대사 조각을 검색해 그대로 보여준다. 디버그 전용 — PERSONA_RETRIEVE_DEBUG_ENABLED(기본 false)가 꺼져 있으면 404. 응답에 원문 조각이 그대로 들어가므로 platform prod overlay에는 이 env를 넣지 않는다.",
        "parameters": [
          {
            "in": "path",
            "name": "persona_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2000
            },
            "description": "질문. 접근 로그에는 남기지 않는다."
          },
          {
            "in": "query",
            "name": "k",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 10,
              "default": 5
            },
            "description": "본문·대사 각각 상위 k개."
          }
        ],
        "responses": {
          "200": {
            "description": "검색 결과 디버그 조회",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrieveResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "415": {
            "$ref": "#/components/responses/Error415"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "사전 발급 정적 토큰 → 서버의 고정 사용자 subject. JWT 아님. 브라우저 메모리만 보관하며 새로고침/탭 종료 후 재입력. 만료·회전은 별도 운영 계약."
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ]
      },
...
    },
...
    "responses": {
      "Error400": {
        "description": "잘못된 문법·ID·cursor·키",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error401": {
        "description": "인증 없음/실패",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error404": {
        "description": "존재하지 않거나 타인 소유",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error409": {
        "description": "상태·revision·키 충돌",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error413": {
        "description": "크기 제한 초과",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Error415": {
        "description": "지원하지 않는 Content-Type/파일 형식",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
...
}